requirements/core.txt
PyYAML>=3.11
boto3>=1.1.3
